+++ lib/src/tokenizer.rs
@@ -1,7 +1,6 @@
 use crate::parsed::ParsedPrefix;
-use crate::Parsed;
+use crate::{Parsed, ParserError};
 use std::collections::{HashMap, HashSet};
-use std::error::Error;
 use std::fmt::Debug;
 use std::marker::PhantomData;
 
@@ -463,27 +462,6 @@ impl<'a, 'b> Iterator for IntoTagsIter<'b, 'a> {
         }
     }
 }
-
-#[derive(Debug, Eq, PartialEq)]
-pub enum ParserError {
-    NoTagKeyEnd,
-    NoTagValueEnd,
-    NoCommand,
-    PrefixWithoutName,
-}
-
-impl std::fmt::Display for ParserError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ParserError::NoTagKeyEnd => write!(f, "Tag Key has no ending '='"),
-            ParserError::NoTagValueEnd => write!(f, "Tag Value has no ending ';' or ' '"),
-            ParserError::NoCommand => write!(f, "Missing command in message"),
-            ParserError::PrefixWithoutName => write!(f, "Prefix has to have name included"),
-        }
-    }
-}
-
-impl Error for ParserError {}
 
 #[derive(Default, Clone)]
 pub struct PartialCfg<'a> {
